requirements.txt
python-telegram-bot==21.6
pillow
openai
requests
